PRSD-703: Move getLandlordHasRegisteredProperties to LandlordService
