+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/controllers/DeregisterLandlordController.kt
@@ -15,14 +15,14 @@ import uk.gov.communities.prsdb.webapp.forms.journeys.LandlordDeregistrationJour
 import uk.gov.communities.prsdb.webapp.forms.journeys.factories.LandlordDeregistrationJourneyFactory
 import uk.gov.communities.prsdb.webapp.forms.steps.DeregisterLandlordStepId
 import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.LandlordDeregistrationCheckUserPropertiesFormModel.Companion.USER_HAS_REGISTERED_PROPERTIES_JOURNEY_DATA_KEY
-import uk.gov.communities.prsdb.webapp.services.LandlordDeregistrationService
+import uk.gov.communities.prsdb.webapp.services.LandlordService
 import java.security.Principal
 
 @Controller
 @RequestMapping("/$DEREGISTER_LANDLORD_JOURNEY_URL")
 class DeregisterLandlordController(
     private val landlordDeregistrationJourneyFactory: LandlordDeregistrationJourneyFactory,
-    private val landlordDeregistrationService: LandlordDeregistrationService,
+    private val landlordService: LandlordService,
 ) {
     @PreAuthorize("hasRole('LANDLORD')")
     @GetMapping("/${CHECK_FOR_REGISTERED_PROPERTIES_PATH_SEGMENT}")
@@ -33,7 +33,7 @@ class DeregisterLandlordController(
         val formData =
             mutableMapOf(
                 USER_HAS_REGISTERED_PROPERTIES_JOURNEY_DATA_KEY
-                    to landlordDeregistrationService.getLandlordHasRegisteredProperties(principal.name),
+                    to landlordService.getLandlordHasRegisteredProperties(principal.name),
             )
 
         return landlordDeregistrationJourneyFactory
