+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/controllers/DeregisterLandlordControllerTests.kt
@@ -17,7 +17,7 @@ import uk.gov.communities.prsdb.webapp.forms.journeys.LandlordDeregistrationJour
 import uk.gov.communities.prsdb.webapp.forms.journeys.factories.LandlordDeregistrationJourneyFactory
 import uk.gov.communities.prsdb.webapp.forms.steps.DeregisterLandlordStepId
 import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.LandlordDeregistrationCheckUserPropertiesFormModel.Companion.USER_HAS_REGISTERED_PROPERTIES_JOURNEY_DATA_KEY
-import uk.gov.communities.prsdb.webapp.services.LandlordDeregistrationService
+import uk.gov.communities.prsdb.webapp.services.LandlordService
 
 @WebMvcTest(DeregisterLandlordController::class)
 class DeregisterLandlordControllerTests(
@@ -27,7 +27,7 @@ class DeregisterLandlordControllerTests(
     private lateinit var landlordDeregistrationJourneyFactory: LandlordDeregistrationJourneyFactory
 
     @MockBean
-    private lateinit var landlordDeregistrationService: LandlordDeregistrationService
+    private lateinit var landlordService: LandlordService
 
     @MockBean
     private lateinit var landlordDeregistrationJourney: LandlordDeregistrationJourney
@@ -56,7 +56,7 @@ class DeregisterLandlordControllerTests(
     fun `checkForRegisteredProperties caches userHasRegisteredProperties then returns a redirect to the are you sure step`() {
         landlordDeregistrationJourney = mock()
         whenever(landlordDeregistrationJourneyFactory.create()).thenReturn(landlordDeregistrationJourney)
-        whenever(landlordDeregistrationService.getLandlordHasRegisteredProperties(anyString())).thenReturn(false)
+        whenever(landlordService.getLandlordHasRegisteredProperties(anyString())).thenReturn(false)
         whenever(
             landlordDeregistrationJourney
                 .completeStep(
